// src/app/shared/models/IPayLeaves.ts
export interface IPayLeaves {
  EmpCode : string;
    ID: number;
   Date: Date;
    From: Date;
    To : Date;
    Days : number;
    Type: string;
    Remarks: string;
    IsSync : boolean;
    RowState: string;
    AddOn: Date
    AddBy: string;
    EditOn: Date;
    EditBy: string
    SyncDate: Date;
   CompanyID: number;
   
  }
  